Move relevant interfaces into a serde directory
