+++ planner/src/common/serde/segment_type.ts
